pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zonny-helper"
version = "0.1.0"
description = "AI-powered CLI for git automation and codebase intelligence"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["cli", "git", "ai", "codebase", "developer-tools", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "pydantic>=2",
    "pydantic-settings>=2",
    "tomli>=2; python_version < '3.11'",
    # tree-sitter is used in Phase 3 (tree build). Listed here as a dependency
    # so the package installs cleanly. Grammar packs are installed in Phase 3.
    "tree-sitter>=0.21",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30"]
openai    = ["openai>=1.30"]
gemini    = ["google-genai>=1.0"]
all       = ["zonny-helper[anthropic,openai,gemini]"]
dev       = [
    "pytest>=8",
    "pytest-snapshot",
    "pytest-cov",
    "ruff",
    "mypy",
    "types-requests",
]

[project.scripts]
zonny = "zonny_helper.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/zonny-helper"
Repository = "https://github.com/yourusername/zonny-helper"
Issues = "https://github.com/yourusername/zonny-helper/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/zonny_helper"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=zonny_helper --cov-report=term-missing"
